% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{addr_family}
\alias{addr_family}
\title{Read the family of an address}
\usage{
addr_family(x)
}
\arguments{
\item{x}{A \code{raddr_address} vector.}
}
\value{
A factor with levels \code{"v4"}, \code{"v6"} and \code{"v6_4in6"}, \code{NA} for missing
addresses.
}
\description{
Read the family of an address
}
\examples{
addr_family(raddr_address(0L, 0L, 0L, 1L, "v4"))

}
