% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{is_raddr_address}
\alias{is_raddr_address}
\title{Test whether an object is a \code{raddr_address}}
\usage{
is_raddr_address(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
Test whether an object is a \code{raddr_address}
}
\examples{
is_raddr_address(raddr_address(0L, 0L, 0L, 1L, "v4"))
is_raddr_address("127.0.0.1")

}
