% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{is_raddr_class}
\alias{is_raddr_class}
\title{Test whether an object is a \code{raddr_class}}
\usage{
is_raddr_class(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
Test whether an object is a \code{raddr_class}
}
\examples{
is_raddr_class(addr_classify(addr_pton("127.0.0.1")))
is_raddr_class("127.0.0.1")

}
