% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{is_raddr_embedding}
\alias{is_raddr_embedding}
\title{Test whether an object is a \code{raddr_embedding}}
\usage{
is_raddr_embedding(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
The elements of the \code{embeddings} column of a \code{\link[=addr_classify]{addr_classify()}} result. There
is no public constructor: these are produced by classification, not built by
hand.
}
\examples{
is_raddr_embedding(addr_embeddings(addr_pton("64:ff9b::a9fe:a9fe"))[[1]])
is_raddr_embedding("169.254.169.254")

}
