% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{is_raddr_parse}
\alias{is_raddr_parse}
\title{Test whether an object is a \code{raddr_parse}}
\usage{
is_raddr_parse(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
Test whether an object is a \code{raddr_parse}
}
\examples{
is_raddr_parse(addr_parse("127.0.0.1"))
is_raddr_parse("127.0.0.1")

}
